// apps/web/src/lib/api/generated/model/transactionOut.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * BrokerOS API
 * OpenAPI spec version: 1.0.0
 */

export interface TransactionOut {
  /** @pattern ^-?\d{1,16}(\.\d{1,4})?$ */
  amount: string;
  bill_reference: string | null;
  created_at: string;
  id: string;
  invoice_id: string | null;
  note: string | null;
  paid_at: string | null;
  payer: string | null;
  payment_id: string | null;
  receipt: string;
  source: string;
  status: string;
}